beacon and other float data have been multiplied by 100 to convert into int16_t . ALl of the contents of sat_health is now of type int16_t

+++ custom_apps/cubus_app/common_functions.h
@@ -45,15 +45,15 @@
 
 typedef struct  {
 	// uint64_t timestamp;
-	float accl_x;
-	float accl_y;
-	float accl_z;
-	float gyro_x;
-	float gyro_y;
-	float gyro_z;
-	float mag_x;
-	float mag_y;
-	float mag_z;
+	int16_t accl_x;
+	int16_t accl_y;
+	int16_t accl_z;
+	int16_t gyro_x;
+	int16_t gyro_y;
+	int16_t gyro_z;
+	int16_t mag_x;
+	int16_t mag_y;
+	int16_t mag_z;
 	int16_t temp_x;
 	int16_t temp_x1;
 	int16_t temp_y;
